# Optimize element-wise operations