Fixes https://github.com/JuliaGPU/CUDA.jl/issues/307

For `lmul!`:

```julia
function main(N=10784)
    println("Knet.jl")
    display(@benchmark(CUDA.@sync(blocking=false, lmul!(2,a)),
                       setup=(a=KnetArray(rand(Float32,$N))),
                       teardown=(Knet.freeKnetPtrCu(a.ptr))
                       ))

    println()

    println("CUDA.jl")
    display(@benchmark(CUDA.@sync(blocking=false, custom_lmul!(2,a)),
                       setup=(a=CuArray(rand(Float32,$N))),
                       teardown=(CUDA.unsafe_free!(a))
                                 ))

    return
end
```

```
Knet.jl
BenchmarkTools.Trial: 
  memory estimate:  288 bytes
  allocs estimate:  12
  --------------
  minimum time:     5.923 μs (0.00% GC)
  median time:      6.115 μs (0.00% GC)
  mean time:        6.142 μs (0.00% GC)
  maximum time:     24.955 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     6

CUDA.jl
BenchmarkTools.Trial: 
  memory estimate:  928 bytes
  allocs estimate:  35
  --------------
  minimum time:     6.281 μs (0.00% GC)
  median time:      6.510 μs (0.00% GC)
  mean time:        6.620 μs (0.83% GC)
  maximum time:     783.503 μs (69.80% GC)
  --------------
  samples:          10000
  evals/sample:     5
```

@denizyuret, do note that for these operations (N=10,784 as you mentioned) actual GPU execution time is only about 600ns, so you should queue up sufficient work (in which case the difference would not have mattered, since the kernels take equal time).